Set up the initial state of the project
In order to get things rolling properly on the first day of Devember,
this day 0 change includes the structure of the project as a whole.

The folder `remote_build_server` contains the stub dotnet core server
project, while the top level will contain the Sublime portion of things.

The remote build server stub is based on https://bit.ly/2i8VUbK but has
had the style altered to match my own with some more clear comments in
it.

The test client is similarly somewhat based on https://bit.ly/2U1sVFd
but modified to be a simple command.

As currently written, the Sublime command can transmit data to the
server and read a response back, which is simple proof of concept that
the two can talk to each other.