Include a stub protocol message class
This includes a stub class to be the common base of all protocol
messages. The general idea is that such a message can be constructed
either by passing keyword arguments to the constructor or by passing a
special `_data` argument to get it to decode a binary block of data.

Similarly, it should be possible to ask an already constructed instance
to encode itself into such a block for transmission.