Include a stub file with the network design
With this day being all about the research and design of how we (most
likely) want the network code to work and be structured, this commit
introduces a file that lays out the various classes and roughly what
their jobs will be.

The idea is to flesh this out over the next few days into a tested
system so that it can be rolled into the plugin.

In contrast, the server component should be easier to structure because
it's a distinct application that is dedicated to networking and already
has an appropriate async layout as it is.