Implement the ProtocolMessaeg class and subclasses
This change uses the previous testbed as a base on which to build our
ProtocolMessage class, which will wrap all of our data transmissions to
and from the remote build host.

This is implemented using the previously outlined design, wherein a
subclass registers itself with the ProtocolMessage base class and can
then be used automagically in decoding a message based on it's internal
type (from the block).

The ProtocolMessage class has some checks in place to ensure that you
don't register messages with the same ID and that you're calling it
with a class that's a subclass of the ProtocolMessage class.