Shift the message classes to their own file
For simplicity now and for potentially future development in the plugin
proper, this shifts the protocol message class and it's subclasses out
into their own file, and imports them back into the main network test
file.

Things still function as they did previously, but now the code is a
lot clearer.

In addition things were cleaned up a little bit in the comment
department.