Implement an initial ConnectionManager class
This fleshes out the ConnectionManager class to provide what may be its
entire API. We can spin up and shut down the network thread, as well as
create and find connections.

Along the way this also has a stubbed out implementation of the network
thread and the connection class; just enough to verify that this code
does what we want.