Implement part of the Connection class
This implements part of the glue of the connection class, which is
responsible for doing all of the grunt work in the networking code.

This code sets up the ability to provide it with messages to transmit
as well as for being able to retrieve messages once we can receive them.

A bit more glue work needs to happen and we can start integrating it
with and developing on the core Network thread responsible for
dispatching calls to the code here.