Shift the network code out for easier testing
This duplicates the existing network code into a test folder to make
Sublime ignore it while we're working.

In addition we've also made some changes to the code as it appeared in
the package proper that makes the ConnectionManager class wholly
responsible for the management of the underlying socket in the
Connection object.