Allow connections to track their connected status
This includes a new boolean field in the connection that tracks whether
or not it is connected to the remote host or not. This defaults to
False, but the first call to the _send() method in which the connection
is not connected, we check to see if the socket is in error or not and
act accordingly.

This will be invoked by the network thread when we select for writing
for the first time, which will be our indication that we are ready to
go.