Include block length prefix bytes in transmissions
This adds in the missing component in our protocol whereby we transmit
exactly how big a block is prior to sending the actual data along.

This also partially implements the send code's ability to track a
partial send by adding an instance variable that tracks the block that
is currently being sent.