Support partial sends on connection objects
This implements the necessary code so that we can handle short sends.

We check how many bytes we actually transmitted to the remote end of the
connection, and if it wasn't the entire message we advance the buffer
according to how much we sent so that we can finish on the next round
through the network thread.