Support partial receives on connection objects
As in the previous commit, this modifies the read code so that it can
handle reading not an entire message at once.

Data we read is buffered and consumed once we know there's enough of it;
first the four bytes that tell us how big the block of data is, and then
the bytes of the actual block itself.