Allow for handling multiple messages in/out
To finalize the basis of the core network code for the client, this
modifies the receive and send code both so that they try to handle
multiple messages at once.

The send code tries to send as many as 10 messages per call to avoid I/O
starvation of other connections, while the receive code tries to receive
as many messages as it can in a single 4k read.

This could likely be enhanced to allow multiple reads as well, but for
now this is good enough to get rolling.