Include Jon Skeet's MiscUtil library
For upcoming code in the server we need to be able to encode and decode
binary data in Network (big endian) format, but it turns out that the
code in .NET always works only in the local endian and requires you to
reverse byte arrays as needed.

Enter this library, which includes (amongst other things) the ability to
do this in a big endian way.