Include a first take on message encode/decode
This includes our first stab at a design for a message system that lets
us encode and decode binary messages into object instances, which will
allow us to better communicate with the client process running inside of
Sublime.

This code needs more cleanup and possibly a bit of a redesign (at the
least every message should maybe be in it's own file or something), but
for now this is good enough.