Shift the build client class to its own file
For clarity of code, this moves the BuildClient code out of the main
testbed code and into it's own implementation file.

Along the way (not done as separate commits because we weren't sure what
we could get away with) the network code is also fully abstracted out of
the main code and into the client code for clarity.