Include a partial message class for decoding data
To make reading of multiple messages (and partial messages) easier in
the receive code in the server, this implements a simple class that
knows how much data it needs to see in order to get a complete message,
and graciously takes just the right amount of data out of the buffer
for any message.

There are also methods to get the partial message to decode the data
for you (via the factory) as well as to indicate if the message is
completely received or not.