Modify the client read code to read binary data
This is a simple modification to our client code that allows it to read
binary data using the new partial message class.

For the time being it just prints messages it receives to the console
and otherwise does nothing; the next changes will allow it to echo the
data back as well, to prove end to end communications are possible.