Modify our message classes to encode their length
In order to make the send code more straightforward in the C# server
code, and also in order to keep the client code working in a similar
manner, the code for encoding a message into a byte array for sending to
the other end of the connection is now responsible for also including
the length prefix.

The decoders still assume that they're being passed just the bytes for
the message without the length prefix, since it's harder to read the
minds of the system to figure out ahead of time how big the block should
be allocated for reading the message data.