Include a third message type for regular messages
This includes a new message type that's like an error message but which
has no error code because it's just a message.

In theory this could be just a single class and we could determine if
this was an error or not based on whether the code was 0, but I like the
idea of having a distinct message for that instead.