Clean up the network code a bit
Now that it's a new day and we're not all tired, this fixes the quick
integration that we did in the last stream to get the network code
working in the plugin in Sublime.

There was also a bug in the network code in which a socket was selected
for reading when it wasn't connected yet, which does indeed return an
error, but in the place that we are set up to expect it.