Include a test implementation of notifications
This adds a simple test notification service to the connection object
so that external code can be notified when an interesting event happens.

Currently this is not as informative as it could be, but it proves that
we can detect when we're connected, disconnected, failed to connect OR
received a message.