Include the beginnings of our file collection
This implements a command wherein we'll be working towards gathering up
all of the files that contribute to a build for transfer to the remote
host.

This is essentially a file walk in all of the folders in the current
project or window (and later as specified in the build instead), using
the same rules that Sublime uses for including and excluding files as
it goes.