Include an initial implementation for filtering
This includes an initial version of code that can filter away files and
folders while walking through the list of project folders, resulting in
just collecting the list of files that should actually appear as a part
of the build.

Currently this is naively pruning away entire portions of the tree based
on whether or not it should recurse into subfolders or not. This works
as long as there is only one child level; otherwise when we hit level
two and beyond, we skip folders that we should probably include because
their parents match the spec even if they don't.