Error check folder entries better
In order to ensure the happiness of people that don't want this package
to blow up in their face like an over inflated balloon, this adds some
error checking to the code that handles path entries.

This means that in particular we want to ensure that in cases where
there is no project file for us to be able to determine where relative
paths are relative to, we instead throw an exception if the path is
relative in this case.

Additionally, we also ensure that theres actually a path defined, since
that is also quite important.