Enable _find_project_files explicit path list
In order to be fully functional we need our command to be able to not
only gather the contents of windows with a project in them or just
windows with a folder, but also explicit folder lists defined as a part
of the build system.

As such this function now takes an optional list of folder entries which
would come from the build system, and falls back to asking for the data
from the project or the window only if that's not provided.