Include an inline example of our project setup
This modifies the sample command such that it takes the exact same setup
as the current project.

This is thus a hard coded version of what would happen if you run the
command in the project window, but it lets us verify what happens while
running in other windows and also to verify what it does as compared to
what Sublime does with files in the sidebar using the same setup.