Fix an issue in file gathering wrt path search
It turns out that I was using os.path.relpath to calculate a relative
path in an incorrect way because that tries to find a path TO another
path and not FROM another path, which is what I was expecting because I
apparently can't read.

If we expect a path to be relative to another path, we can join them
together and use abspath() to make the path canonical.