Implement a method to get file details
This takes a root path and the name of a file in that path and returns
back an object that gives you the local name of the file, the last
modification time, and the SHA1 hash of the contents of that file.

This is to be used in a future commit to store information about this
file and it's contents for sharing with the server.