Add new handling to the case of a single file
In the case where there are no folders, the only thing we transmit back
and forth is a single file. In that case we need to modify the existing
code because it was just adding the file name to the list, but now it
needs to split the filename into a base path and a file and it needs to
get file details for that file.