Make SHA1 hashing optional for optimization
Although when we first fetch a list of file details we definitely want
to get the SHA1 contents hash of all of the build files, at the same
time we also want to get file lists later and only calculate the hash
when we know they're probably different (as determined by their last
modification time).

This adds the appropriate arguments with the appropriate boolean default
values to allow this to calculate hashes by default but also not
calculate them when needed, which is a nice time optimization.