Implement disparate folder coalescing
This implements a function that can take a dictionary of build files and
coalesce any common folders away. For example if there are two folders
and one of them has an absolute path that is wholly contained inside the
second, then the items from the second are moved into it and their
relative names are changed.

This essentially cuts down as much as possible the list of paths that
are a part of the build to only the most essential, and reconstructs the
on disk layout in cases where the side bar has purposefully split a
project folder into multiple top level entries.