Implement a file set diff function
This is the last step in being able to calculate deltas between files we
and the server are sharing with each other. Given the output of a file
scan operation (which should have SHA1 hash information in it), we can
create a folder that indicates, for each folder in the build, what files
need to be added, removed or updated to bring things into sync.

Note that for our test purposes we have captured the contents of the
remote build server as of tonight, which includes the hashes it had,
and then modified some of the entries to test the diff.

As such, come back here and punch me when you chase down a bug that
turns out to be caused by forgetting that the files on disk changed
while you were developing them. Sorry!