Include our dotnet_tests project
Much like we have a Python directory (net_test) for doing some Python
work without messing with the core code in the package, this adds a
similar directory for working with C# code in the same way.

Currently this shows how we can use JSON.net to decode and encode some
JSON data.