Include a new extension for fixed length strings
This includes a C# extension method for strings that does the work of
ensuring that they are an exact width (trimming or padding as needed)
and then converting them to a byte[] array using UTF-8 encoding.

This will be used in our messages whenever we want to transmit a fixed
length string.