Use the new introduction message semi-properly
As a part of our protocol, the first thing that should happen once a
connection is made is that we transmit an introduction message to tell
the server who we are.

We now do that as soon as a connection is made, transmitting a user
name and password and letting the class come up with its own platform
and host name.

In addition whenever we receive a message we just display it directly
instead of assuming it has a "msg" field, so that we can verify that
the server is echoing our Introduction message back properly.