Introduce client message dispatching
As an expedient for upcoming changes, this adds in a new Dispatch method
in a client, which it will call whenever it receives a message.

There is now a partial class that uses this to echo the way we always
have up until now, the code is now more ready to continue with what we
are going to do.