Include the ability to close after sending a msg
This includes the core functionality that allows any particular message
to be flagged in a way that tells the sending code that once the send is
complete, we should close the connection to the client.

This is done so that we can control what the last message we transmit is
while still making sure that it's us and not the client that's in charge
of ending connections.

In particular we might want to close the connection after reporting an
error OR to graciously hang up with a "Goodbye" type message, but we
need to make sure that we do it in a smart way because we don't know at
any given point when a particular message will actually be sent.

Here we make the message have a flag that indicates whether the
connection should be closed or not, and the send code closes the socket
once that message is sent if the flag is set.