Allow server Introduction parsing to trim nulls
The server was not properly trimming the null bytes off of the strings
that it receives in Introduction messages, which will no doubt be an
issue going forward.

To that end it has been repaired by way of a new global static method
that can decode bytes as well as doing the trim.