Include an Exclusions class in the C# TestBed
This includes a class that wraps the functionality of excluding a file
or folder while gathering files.

The implementation here is currently case sensitive when it probably
should not be (or more specifically it should be based on the operating
system in use).

This also currently uses simple compares to exclude folders and a regex
to match files. While implementing this I discovered that there is a
library on NuGet that claims to support globbing, though it seems like
it can only use it to scan files and not match them, so for the time
being I've ignored that in favor of this as a more simplistic solution.