Include a function to scan a folder for files
Although somewhat short, this is the product of a long series of
iterations which culminate in something that works similarly to how the
similar code in the Python base works.

This scans folders using recursion so that when we exclude a path we can
reject all files inside of it without having to scan into them first.

This needs the exclusions class from the previous commit to actually
make the decisions