Ensure that we always listen on IPv4 addresses
For local testing purposes, this ensures that we always select the IPv4
local address (if any) to listen for incoming connections on.

This could probably be much better, but this is good enough for now so
that I don't need to keep flip-flopping the code as I move between
machines.