Provide the global configuration to the client obj
This passes the loaded configuration information to each new client as
their connections are accepted, so that they will also have access to
the configuration information.

In the future this will have to be thread-safe and guarded with a mutex
of some sort, but for right now we're not going to worry about that
because we're only ever going to have a single client at the moment
anyway.