Finish the initial client handshake
This ties all of the recent commits together by having the server use
the information from the introduction message to actually find the user
that was provided, and either error the connection out or allow it based
on the credentials provided.

Once done, if the connection is left around the client will have a
record of it's specific configuration information (which for the moment
is not a lot at all).