Include a command to prompt for a password
This command will be part of a command chain and used in places where
we need to ask what the password for a connection should be.

The command assumes that it will get all arguments that would be needed
by whatever command calls it, and that those arguments include the
username and host name being connected to, and prompts the user to input
a password.

Once entered, the build command is executed with all of the arguments
that we were originally given, plus the password.

Whoops