Implement a command to prompt for a connection
This implements a second command that will be used as a part of our
ultimate command chain.

Here we again assume we have all arguments that whoever called us might
need, and we show a list of all of the configured build hosts and let
the user pick one.

If the user picks one that has a password, the build command is
immediately executed using the host, port, user and password from the
selected host.

When there is no password present, the command instead executes the last
added command, which prompts the user for the password to use.