Fix a bug in sending messages that slipped through
The previous commits for Christmas were some of those commits wherein we
refactor code but don't spend a lot of time stopping and restarting
Sublime because it's Christmas Day and we're not as focused on coding
during our few sick-free waking hours.

In any case, although this worked yesterday, with a fresh install it
no longer does because there's not a version of the class loaded that
doesn't take a hostname and port.