Make the gathering command public
Now that the file gathering code has been tested, it should actually be
public in it's module file instead of private, since it's one of the two
reasons for that particular file to exist in the first place.

The test code will remain but will be shifted out into it's own file in
the next commit to clean things up a bit.