Allow the server to listen on the localhost IP
In order to facilitate testing on a variety of machines (and now that
I'm at a nicer keyboard for developing on) this adds a new configuration
element for the server that indicates if it should attempt to listen on
the localhost IP address or one of it's other hostnames instead.

This is preferable to listening by name directly for cross testing. Also
this shows how the config can be altered later to specify the listening
address instead of just assuming the first IPv4 address found.