Include a new message type, SetBuild
This message type will be the first message sent in any build operation
to allow the client to communicate to the server what it is going to try
to build.

The idea is that the server will use this to figure out where and how it
should stash files, and both ends will gather the file lists from those
locations.