Set up the client to use the new message
Replacing the old placeholder of prompting the user for a message and
then sending it to be echoed, we now respond to a build request by
transmitting away the folder list that represents the current build.

Technically this is still not the build per set but our sample command
pretending it's building the files in the current project, since for a
little bit yet we need to still be able to build our code while we're
working on it without constantly swapping the setting around.