Turn off the client side notification logging
Now that we're relatively assured that the communications channel
between the client and the server is working, this removes the longer
logging of every client side notification in favor of just displaying
the messages received.

This will have to be updated in the future so that we can actually
handle the incoming messages more appropriately, but this is good
enough for the time being.