Include some extra processing functions
This includes some extra handling functions for messages and better
documents the ones that exist.

Newly added are the ability to echo a message back to the other end
(for testing purposes) and the ability to generate a protocol violation
message along with a reason why. This is used as an indication that the
client has done something wrong, stating what the unexpected message was
and why it was rejected.