Clarify dispatching a little bit
Using the recently added handlers, this clarifies the handling a little
bit by ensuring that we issue protocol violations in the appropriate
places.

We also redact the client's ability to transmit textual and error
messages because those are messages meant for us to convey information
and not the other way around.