Enhance the SetBuild message to include an ID
In order to make life easier in the server, the SetBuild message has
been enhanced so that it also takes a unique build ID that represents
this build and no other.

The uniqueness constraint is only required for machines on the same host
and OS and is implemented by sorting the list of folders into a known
state and calculating a SHA1 hash of the folders. Thus the hash will
remain the same so long as the same folders are being used to build.

The sample build command uses the folder list to generate the build ID
for us.