Enhance the server end SetBuild message
Although the server doesn't send these messages directly, it does need
to be able to encode them properly since we currently echo them back.

The client version of the message, for expedience here at the end of
Devember, just includes the build_id as one of the folders that it
transmits, so here on the server we replicate that logic when we both
decode and re-encode later on.

The server has no facility for coming up with a build ID; this has to
instead be done on the client, since that's where the build process
originates.