Include code to absolute-ize a relative cache path
The configured cache path (which was previously hard coded for my
windows laptop but now is relative and platform agnostic) has new code
handling at startup that determines if the path is relative or not and
turns it into an absolute path if needed.

This is actually a member in the config object, but not one which gets
persisted to and from the configuration; it's just a runtime constant.