Include nascent handling for SetBuild message
Instead of echoing the message to the remote end of the connection, we
now capture the current build ID and folders, and map the folders to
locations in our remote cache.

This sets us up to be able to receive files from the client and drop
them in the appropriate location.

Note that currently the folder created in the cache has the same top
level component as the configured build. Perhaps this should be a more
deep path (for relative builds), but for the time being this is the
most expedient.