Clean up client side display a bit
Since we're starting to get into the meat of actual communications now,
this refactors the code for displaying received messages so that it is
formatted more "correctly".

Work still needs to happen here, in particular for the information that
is conveyed by the underlying network thread, and we could set up a
syntax to make it easier to read, but this is good enough for now.