Make SendError a varargs function
Similar to the previous commit (but split apart to make the diffs a bit
nicer and for verification that things still work), this makes the
SendError function take a varargs format string for more natural calls.

This changes the call point signature a little and requires that the
flag for whether the connection should close after sending appear before
the format string so it can be properly distinguished from format args.