Implement a new Acknowledge message
This is a new message that is meant to be paired with textual returns
from the server to the client to allow the client to know that the
server has received the message they sent.

This allows our protocol to be driven by server responses while still
providing informational text messages without requiring the client to do
something silly like try to parse the text to know if it should proceed.

The Acknowledge message has the capacity for both a positive and a
negative acknowledgment, should that be needed. Generally I'm
visualizing that any time the server wants to NACK a message it would
be generating an error message and closing the connection, but this
opens up future possibilities.

Like the textual messages, this message type is for server usage only.