Shift the message decode into the dispatcher
For an upcoming error handling change, this shifts the location at which
the raw data is converted into a message for dispatching into the
actual dispatch code from the reading code.

This ensures that we can easily catch any decode errors as well as any
handling errors all in one place without disrupting the reading code.