Ditch the enum library, at least for now
I have been having bizarre intermittent problems for the last week or
two in which sometimes the client did not seem to receive any messages
even though the server thinks that it sent them.

I finally spent a bunch of time tracing through network code tonight to
see if perhaps I had a weird edge case where a message was not being
fully considered received in some cases.

It turns out that the culprit is that sometimes the Notification object
that was in use in the network code was different than the one seen by
the command (but not always, and for reasons I was unable to determine).

This causes the handling to totally fail because "Notification.MESSAGE"
is two different values.

Since we're nearing the end of Devember and I don't have time to fiddle
faddle with that sort of nonsense, I've dropped the enum code completely
in favor of a simple class that just has numeric defines.

Presumably no matter if the object ends up being different values, the
value 1 is still always going to be the value 1, and things will still
work.