Include a FileContent message
This includes a new message in both the client and in the server that
can transmit the contents of a file.

This is done by providing the remote folder name, the relative name of
the file inside of that folder, and then the bytes of the file from the
disk.

In the future this should be a partial message type that allows us to
send the data in chunks instead of all at once, but this is good enough
for our purposes here.

For expedience, only the client message loads the data from disk when a
message is created; the server just uses stub data. It's a phase two
operation for us to have the server transmit files back to the client,
so we don't need to worry about that here.