Make the client code wait for an ack before send
As a proof of concept tweak to the client, now it prepares the file
content message right away but doesn't actually transmit it to the other
end until the server Ack's (positively) the SetBuild message.

On the whole visually this doesn't appear any different than any other
run, but behind the scenes it's "smarter".