Include notifications for send and receive errors
This includes a couple of new notifications that indicate when a
connection is about to close due to a send or receive error.

Currently this doesn't indicate what the error actually was, but I don't
know how useful that is in the general case and it's easy enough to add
a parameter to the notifications later if needed.