Handle notifications in the client code
This handles all of the possible notifications in the client code, which
makes the client ultimately responsible for all of the panel output that
gets generated, since the network code no longer does so.

This also solves a problem that was introduced with the CONNECTING
notification, since the code here blindly drops the connection when
a notification happens to a connection which is not connected, which
made the code decide to drop the connection as soon as the first
notification happened.