Include more proper handling of notifications
This makes the notification code "more proper". It can still be cleaner
(for example, maybe we want an object or some sort of dispatcher) but
for our immediate purposes, it's OK.

This handles acks properly, so now the build is only triggered once the
login succeeds, instead of logging in and immediately starting the build
as we did before.

The file transmission still happens on receipt of the set build ack as
it did before (but it looks nicer now).