Have the client send all files at build start
This can be made smarter later, but for now for every build every file
is transmitted to the remote server.

To do this we capture a list of all files when the build starts and then
transmit them one at a time to the remote end. Once the last file has
been acknowledged, we can go ahead and actually start the build (though
currently that does not yet happen).