Have the server stop echoing file contents back
In the first step towards having the server actually handle the messages
with file content, instead of echoing them back we just acknowledge that
we got them.

The client handling for receiving messages has been disabled for now
since we don't expect to get any (and thus we want them to be unhandled)
but this means that we need to do our own logging as we send messages
so that we can track that something is happening.