Make the server fully handle our file messages
The server now writes the contents of files that it gets into the local
cache directory, which ends up with the server having a mirrored copy of
the data from the client.

This is the last of the prep before we can start the final step, which
is to try to transmit a command to the server to have it execute the
build.