Throw a quick syntax at the output panel
This is extremely crude and uses dumb scopes, but is just here to make
the output in the panel a little bit more colorful and easier to read.

This is very much an example of "never do this" if you're working on an
actual package, but as Weird Al says, it's "good enough for now".