Keep msbuild's nosy face out of the build cache
For our local testing we're distributing the whole of the project folder
to the remote build server, which is also us.

This has the side effect of seriously cheesing off msbuild, which finds
duplicate files and then refuses to compile them because they contain
classes that already exist.

This tells msbuild to keep it's face out of the cache folder.

Note that intermittently PackageDev sometimes gets mad about the syntax
or settings being empty if it happens to notice that they exist before
the content gets fully written. I'm not sure what a good way around
that is (short of not having the cache bundled in the project for
clean testing across multiple machines), but it's less invasive than
msbuild.