Set syntax setting instead of using assign_syntax
The `view.assign_syntax()` API endpoint is the "correct" way to give a
view a syntax definition. It causes internal state on the view to change
such as ensuring that the syntax specific settings get applied properly.

Anecdotally, views created via `window.create_output_panel()` are
considered to be widgets and use the Widget context menu. However once
you `assign_syntax()` a setting onto them, they lose their widget status
and start using the regular context menu.

Another way to give a view a syntax is to set the `syntax` setting on
them directly. This changes how the syntax highlighting works, but it
doesn't do the internal state changes like applying syntax specific
settings.

This has the side effect of keeping the panel as a widget. Currently
that's what we want (the regular context menu is too busy and doesn't
contain anything applicable to the panel), so for now that's how we are
going to handle this.

The other way to go would be to use `assign_syntax()` and put our
single context menu item for clearing the output panel into the regular
context menu.