Set up some default build arguments
For expedience we're going to test out the last stages of the prototype
by running the command as-is instead of as a standard build command,
since that requires a bit more plumbing than we want to go to for this
last day of devember.

To that end, we have set up build folder that points to the packaged
test_program files (this requires that the package not be installed
packed) and a shell_cmd that will execute it.

We also set the build to run using those folders instead of the window
folders.