Include some message types for executing the build
This includes three new messages; one for transmitting a request to
actually launch a build, one for the server to send results back to us
and one for the server to transmit the exit code once the build has
completed.

These messages are implemented in both the server and in the client,
though the code that tests them (by echoing them back to the client)
is not checked in while we finish the implementation.