Include the initial execution code
This adds in all of the required handling to actually execute the task
and transmit the information back to the client code.

This more or less works, but I believe that it has exposed an inherent
threading bug in our send code, in which it is querying and/or modifying
some instance variables from the external async thread.

The fix for that is next, and then things should hopefully do what we
want.