Include a mutex around the send code
This resolves the issue that we just uncovered in which there is a race
condition if multiple threads try to send messages at the same time.

In particular, even though the queue we use is thread safe, the send
code uses some member variables that were not previously protected in
order to determine if it should try to start a send or not.

This is now protected with a mutex so that a thread is temporarily
blocked if it tries to send while the send code is getting read to
actually perform a send.